pyproject.toml
[build-system]
requires = [
    "setuptools==80.9.0",
    "setuptools_scm==8.3.1",
]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "beeware"
description = "A metapackage to install the full BeeWare suite of tools."
readme = "README.rst"
requires-python = ">= 3.8"
license.text = "New BSD"
authors = [
    {name="Russell Keith-Magee", email="russell@keith-magee.com"},
]
maintainers = [
    {name="BeeWare Team", email="team@beeware.org"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development",
]
dependencies = [
    "briefcase",
]

[project.urls]
Homepage = "https://beeware.org"
Funding = "https://beeware.org/contributing/membership/"
Documentation = "https://docs.beeware.org/"
Source = "https://github.com/beeware"

[project.optional-dependencies]
dev = [
    # Pre-commit 3.6.0 deprecated support for Python 3.8
    "pre-commit == 3.5.0 ; python_version < '3.9'",
    "pre-commit == 4.2.0 ; python_version >= '3.9'",
    "pytest == 8.4.1",
    "pytest-tldr == 0.2.5",
    "setuptools_scm == 8.3.1",
    "tox == 4.27.0",
]
# Docs are always built on a specific Python version; see RTD and tox config files.
docs = [
    "black",
    "codespell",
    "markdown-checker",
    "mkdocs",  # Is this needed?
    "mkdocs-autorefs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-literate-nav",
    "mkdocs-material",
    "mkdocs-macros-plugin",
    "mkdocs-spellcheck",
    "mkdocs-table-reader-plugin",
    "mkdocstrings-python",
    "mistletoe",
    "pymdown-extensions",
    "python-Levenshtein",
    "symspellpy",  # Not used, but spellcheck fails without it. Bug report filed.
    "translate-toolkit @ git+https://github.com/kattni/translate@md2po-line-numbers",
]

[tool.codespell]
skip = ".git,*.pdf,*.svg"
# the way to make case sensitive skips of words etc
ignore-regex = "\bNd\b"
ignore-words-list = "re-use"

[tool.isort]
profile = "black"
split_on_trailing_comma = true
combine_as_imports = true

[tool.setuptools]
py-modules = []

[tool.setuptools_scm]
# To enable SCM versioning, we need an empty tool configuration for setuptools_scm

[tool.check-wheel-contents]
# WOO7: wheel library is empty
# W008: wheel is empty
ignore = "W007,W008"
